firmware/sys/storage: Added supports to del registers an rwp bit set
