+++ firmware/sys/storage/Makefile
@@ -1,3 +1,6 @@
 SRC = storage.c
+SRC += storage_register.c
+SRC += storage_internal.c
+
 
 include $(RIOTBASE)/Makefile.base
